feat: add AddToFavoritesBtn component to handle favorite movie functionality in MovieHero
